+++ tests/unit/components/test_accordion.py
@@ -1,0 +1,11 @@
+"""Tests for the accordion component."""
+
+from tests.unit.components.test_template_tags import TemplateTagsTestCase
+
+# TODO: Implement tests for Accordion component  # noqa: TD002, TD003
+
+
+class TestAccordion(TemplateTagsTestCase):
+    """Test suite for the accordion component."""
+
+    pass